bitstream-extraction_git.bb: Handle empty bitstream
This patch fixes the issue when you have HDF without bitstream.  Having
bitstream in BOOT.bin is enabled by default, if the bitstream is missing
it will continue to create BOOT.bin without it.

+++ recipes-bit/bitstream/bitstream-extraction_git.bb
@@ -15,7 +15,11 @@ do_compile[noexec] = "1"
 
 do_deploy() {
         install -d ${DEPLOYDIR}
-        install -m 0644 ${XSCTH_WS}/${XSCTH_PROJ}_hwproj/*.bit ${DEPLOYDIR}/download-${MACHINE}.bit
+        if [ -e ${XSCTH_WS}/${XSCTH_PROJ}_hwproj/*.bit ]; then
+                install -m 0644 ${XSCTH_WS}/${XSCTH_PROJ}_hwproj/*.bit ${DEPLOYDIR}/download-${MACHINE}.bit
+        else
+                touch ${DEPLOYDIR}/download-${MACHINE}.bit
+        fi
 }
 
 addtask do_deploy after do_install
